もちっと見直し

+++ common/.zshrc
@@ -182,9 +182,13 @@ if [ -S $agent ]; then
 	if [ "x$SSH_AUTH_SOCK" != "x$agent" ]; then
 		export SSH_AUTH_SOCK_ORIG=`readlink $agent`
 		if [ "x$SSH_AUTH_SOCK" != "x$SSH_AUTH_SOCK_ORIG" ]; then
-		    rm -f $agent
-		    ln -snf "$SSH_AUTH_SOCK" $agent
-		    echo "relink $agent from $SSH_AUTH_SOCK_ORIG to $SSH_AUTH_SOCK"
+		    if [ -S $SSH_AUTH_SOCK ]; then
+			rm -f $agent
+			ln -snf "$SSH_AUTH_SOCK" $agent
+			echo "relink $agent from $SSH_AUTH_SOCK_ORIG to $SSH_AUTH_SOCK"
+		    else
+			echo "reuse link $agent from $SSH_AUTH_SOCK to $SSH_AUTH_SOCK_ORIG"
+		    fi
 		else
 		    echo "reuse link $agent to $SSH_AUTH_SOCK_ORIG"
 		fi
